requirements.txt
python-dotenv==1.0.0
transformers==4.31.0